accept either .mp4 or dir of .jpgs as input to calibrate_checkerboard()

+++ calibrate_checkerboard.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
 
 """
-Calibrates and undistorts images from a calibration checkerboard video.
-The rows and corners of the checkerboard are interchangeable. 
+Undistorts videos by calibrating a checkerboard .mp4 video or a 
+folder of checkerboard .jpg images. 
+The number of internal corners on the checkerboard's rows and columns 
+are interchangeable. 
 """
 
 import subprocess
@@ -83,7 +85,7 @@ def convert_vid_to_jpgs(vid, framerate, backend="opencv"):
             subprocess.run(args, cwd=jpgs_dir)
 
 
-def calibrate_checkerboard(board_vid, m_corners, n_corners, framerate=30, do_debug=False):
+def calibrate_checkerboard(board_vid, m_corners, n_corners, framerate=30, do_debug=True):
 
     """
     Finds internal corners of checkerboards and saves them from a folder of .jpgs.
@@ -91,21 +93,23 @@ def calibrate_checkerboard(board_vid, m_corners, n_corners, framerate=30, do_deb
 
     Parameters:
     -----------
-
     board_vid (str): Path to .mp4 checkerboard video or path to a folder containing 
         checkerboard .jpgs.
     m_corners (int): Number of internal corners along the rows of the checkerboard
     n_corners (int): Number of internal corners along the columns of the checkerboard
     framerate (int): Framerate with which `board_vid` was recorded
-    do_debut (bool): 
+    do_debug (bool): If True, will show a live feed of the labelled checkerboards, and
+        will save a directory of the labelled checkerboard .jpgs. Default is True. 
 
     Returns:
     --------
     ret, cam_mtx, dist, r_vecs, t_vecs from cv2.calibrateCamera() 
     """
     
-    assert(splitext(board_vid)[1] == ".mp4"), "`board_vid` must be an '.mp4' file!"
     assert(basename(board_vid) != "checkerboards.mp4"), "Rename 'checkerboards.mp4' to something else!"
+
+    if Path(board_vid).is_file():
+        assert(splitext(board_vid)[1] == ".mp4"), "`board_vid` must be an '.mp4' file!"
 
     output_vid = path.join(dirname(board_vid), "checkerboards.mp4")
 
@@ -153,40 +157,83 @@ def calibrate_checkerboard(board_vid, m_corners, n_corners, framerate=30, do_deb
         img_points = [] # 2d points in image plane
 
         i = 0
-        cap = cv2.VideoCapture(board_vid)
-        while (cap.isOpened()):
-            
-            ret, frame = cap.read()
-            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-            # Find the chess board corners:
-            ret, corners = cv2.findChessboardCorners(gray, (m_corners, n_corners), None)
+        if Path(board_vid).is_file():
 
-            # If found, add object points, image points (after refining them):
-            if ret == True:
+            cap = cv2.VideoCapture(board_vid)
+            while (cap.isOpened()):
                 
-                obj_points.append(obj_p)
+                ret, frame = cap.read()
+                gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-                # This method increases the accuracy of the identified corners:
-                better_corners = cv2.cornerSubPix(gray, corners, (11,11), (-1,-1), criteria)
-                img_points.append(better_corners)
+                # Find the chess board corners:
+                ret, corners = cv2.findChessboardCorners(gray, (m_corners, n_corners), None)
 
-                # Draw and display the corners:
-                img = cv2.drawChessboardCorners(frame, (m_corners, n_corners), better_corners, ret)
-                
-                # Save to video:
-                out.write(img)
+                # If found, add object points, image points (after refining them):
+                if ret == True:
+                    
+                    obj_points.append(obj_p)
 
-                if do_debug:
+                    # This method increases the accuracy of the identified corners:
+                    better_corners = cv2.cornerSubPix(gray, corners, (11,11), (-1,-1), criteria)
+                    img_points.append(better_corners)
 
-                    cv2.imwrite(path.join(boards_dir, f"frame_{i:08d}.jpg"), img)
-                    cv2.imshow("checkerboard detected ...", img) 
-                    if cv2.waitKey(1) & 0xFF == ord("q"):
-                        break
+                    # Draw and display the corners:
+                    img = cv2.drawChessboardCorners(frame, (m_corners, n_corners), better_corners, ret)
+                    
+                    # Save to video:
+                    out.write(img)
 
-                print(f"found {i+1} checkerboards so far ...")
-                cv2.waitKey(1) 
-                i += 1
+                    if do_debug:
+
+                        cv2.imwrite(path.join(boards_dir, f"frame_{i:08d}.jpg"), img)
+                        cv2.imshow("checkerboard detected ...", img) 
+                        if cv2.waitKey(1) & 0xFF == ord("q"):
+                            break
+
+                    print(f"found {i+1} checkerboards so far ...")
+                    cv2.waitKey(1) 
+                    i += 1
+
+                    cap.release()
+
+        elif Path(board_vid).is_dir():
+
+            jpgs = [str(path.absolute()) for path in Path(board_vid).rglob("*.jpg")]
+
+            for jpg in jpgs:
+
+                img = cv2.imread(jpg)
+                gray = cv2.cvtColor(img,cv2.COLOR_BGR2GRAY)
+
+                # Find the chess board corners:
+                ret, corners = cv2.findChessboardCorners(gray, (m_corners, n_corners),None)
+
+                # If found, add object points, image points (after refining them):
+                if ret == True:
+                    
+                    obj_points.append(obj_p)
+
+                    # This method increases the accuracy of the identified corners:
+                    better_corners = cv2.cornerSubPix(gray, corners, (11,11), (-1,-1), criteria)
+                    img_points.append(better_corners)
+
+                    # Draw and display the corners:
+                    img = cv2.drawChessboardCorners(img, (m_corners, n_corners), better_corners, ret)
+                    
+                    # Save to video:
+                    out.write(img)
+
+                    if do_debug:
+
+                        cv2.imwrite(path.join(boards_dir, f"frame_{i:08d}.jpg"), img)
+                        cv2.imshow("checkerboard detected ...", img) 
+                        if cv2.waitKey(1) & 0xFF == ord("q"):
+                            break
+
+                    print(f"found {i+1} checkerboards from {len(jpgs)} frames")
+                    cv2.waitKey(1) # can't be too low, 500 seems safe if using imshow()
+                    i += 1
         
         out.release
         cv2.destroyAllWindows()
@@ -252,6 +299,9 @@ def main():
         help="Number of internal rows along the corners of the checkerboard")
     parser.add_argument("vid_to_undistort",
         help="Path to the target video for undistortion")
+    parser.add_argument("--debug", "-d", action="store_true", 
+        help="Show a live feed of the labelled checkerboards, and save a \
+            directory of the labelled checkerboards as .jpgs.")
     args = parser.parse_args()
 
     board_vid = expanduser(args.board_vid)
